Chore: rename CloudWatch mock file
