+++ public/app/plugins/datasource/cloudwatch/__mocks__/CloudWatchVariables.ts
